Split out command module.
